Implemented basic user input / hint drawing

+++ src/main/java/at/primetshofer/model/Point.java
@@ -1,26 +1,26 @@
 package at.primetshofer.model;
 
 public class Point {
-    private int x, y;
+    private double x, y;
 
-    public Point(int x, int y) {
+    public Point(double x, double y) {
         this.x = x;
         this.y = y;
     }
 
-    public void setX(int x) {
+    public void setX(double x) {
         this.x = x;
     }
 
-    public void setY(int y) {
+    public void setY(double y) {
         this.y = y;
     }
 
-    public int getX() {
+    public double getX() {
         return this.x;
     }
 
-    public int getY() {
+    public double getY() {
         return this.y;
     }
 }
